do not run editstore migrations in production/staging -- they will already be ready

+++ db/migrate/20130808001054_update_fields_in_editstore.rb
@@ -1,14 +1,13 @@
 class UpdateFieldsInEditstore < ActiveRecord::Migration
   def up
-    @connection=Editstore::Connection.connection
-    project=Editstore::Project.where(:name=>'Revs').first
-    Editstore::Field.create(:name=>'city_sections_ssi',:project_id=>project.id)
-    Editstore::Field.create(:name=>'states_ssi',:project_id=>project.id)
-    Editstore::Field.create(:name=>'countries_ssi',:project_id=>project.id)
-    Editstore::Field.create(:name=>'cities_ssi',:project_id=>project.id)
-    Editstore::Field.where(:name=>'location_ssi').first.destroy
-  end
-
-  def down
+    if Editstore.run_migrations?
+      @connection=Editstore::Connection.connection
+      project=Editstore::Project.where(:name=>'Revs').first
+      Editstore::Field.create(:name=>'city_sections_ssi',:project_id=>project.id)
+      Editstore::Field.create(:name=>'states_ssi',:project_id=>project.id)
+      Editstore::Field.create(:name=>'countries_ssi',:project_id=>project.id)
+      Editstore::Field.create(:name=>'cities_ssi',:project_id=>project.id)
+      Editstore::Field.where(:name=>'location_ssi').first.destroy
+    end
   end
 end
